cocotb/requirements.txt
cocotb==1.8.0
cocotb-bus==0.2.1
cocotbext-axi==0.1.24
find-libpython==0.3.0
